+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## ?.?.?
+
+* Add new SDK `WebhookClient` for managing webhook tokens with okdata-permission-api.
+
 ## 0.8.1
 
 * Fix handling of Keycloak tokens that don't contain a refresh token.